+++ packages/api/package.json
@@ -78,6 +78,7 @@
     "node-fetch": "2.7.0",
     "keyv": "^5.3.2",
     "zod": "^3.22.4",
-    "tiktoken": "^1.0.15"
+    "tiktoken": "^1.0.15",
+    "axios": "^1.8.2"
   }
 }
